Added security to products.view

+++ products/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, reverse, get_object_or_404, redirect
 from django.contrib import messages
 from django.db.models import Q
-
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductForm
 from django.db.models.functions import Lower
@@ -10,6 +10,9 @@ from django.db.models.functions import Lower
 # Create your views here.
 def all_products(request):
     """ A view to show all products, including sorting and search queries """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse,('home'))
 
     products = Product.objects.all()
     query = None
@@ -71,6 +74,7 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
     
+@login_required
 def add_product(request):
     """ Add a product to the store """
     if request.method == 'POST':
@@ -91,8 +95,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse,('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -114,8 +123,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse,('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
